// src/libs/enums/product.enum.ts
export enum ProductUnit {
  KG = "KG",
  LITER = "LITER",
  PIECE = "PIECE",
}

export enum ProductStatus {
  PAUSE = "PAUSE",
  PROCESS = "PROCESS",
  DELETE = "DELETE",
}

export enum ProductCollection {
  FRUIT_VEGETABLE = "FRUIT_VEGETABLE",
  MEAT = "MEAT",
  DAIRY = "DAIRY",
  BAKERY = "BAKERY",
  BEVERAGE = "BEVERAGE",
  SNACK = "SNACK",
  OTHER = "OTHER",
}